Tests passing except for error in get_data summary stats
