+++ dashboard/fixtures/10_extractedchemical.yaml
@@ -1,4 +1,15 @@
 - model: dashboard.extractedchemical
+  pk: 1
+  fields: {created_at: null, updated_at: null, extracted_text: 156051, raw_cas: 120-47-8,
+    raw_chem_name: ethylparaben, raw_min_comp: '.0047', raw_max_comp: '.0074', unit_type: 1,
+    report_funcuse: tubing, weight_fraction_type: 1, ingredient_rank: null, raw_central_comp: ''}
+- model: dashboard.extractedchemical
+  pk: 3
+  fields: {created_at: null, updated_at: null, extracted_text: 156051, raw_cas: 80-05-7,
+    raw_chem_name: bisphenol a, raw_min_comp: '4', raw_max_comp: '7', unit_type: 2,
+    report_funcuse: cleaning, weight_fraction_type: 1, ingredient_rank: null, raw_central_comp: ''}
+
+- model: dashboard.extractedchemical
   pk: 5
   fields: {created_at: null, updated_at: null, extracted_text: 7, raw_cas: 84-74-2,
     raw_chem_name: dibutyl phthalate, raw_min_comp: '4', raw_max_comp: '7', unit_type: 1,
